Added throttling

+++ src/main/scala/TwitchManager.scala
@@ -26,7 +26,13 @@ case object FullySubscribed extends TwitchState
 
 case class TwitchUser(name: String, lastAnnounced: String, subscribers: List[Subscriber]) {
   def addSubscriber(sub: Subscriber): TwitchUser = {
-    copy(subscribers = sub :: subscribers)
+    if (subscribers.find(p => p.channel == sub.channel).isDefined) {
+      sub.actor ! AlreadySubscribed(name)
+      this
+    } else {
+      copy(subscribers = sub :: subscribers)
+    }
+
   }
 
   def updateLastAnnounced(lastAnnouncedShow: String): TwitchUser = {
@@ -53,33 +59,38 @@ class TwitchManager extends Actor with FSM[TwitchState, TwitchData] {
 
   when(Open) {
     case Event(PollTwitch, followed: Followed) =>
-      val commaList = followed.twitchUsers.foldLeft("") {
-        case (str, (key, value)) =>
-          str + key + ","
-      }.dropRight(1)
-      val url = "http://api.justin.tv/api/stream/list.json?channel="
-      println("url is" + url)
-      val json = Source.fromURL(url + commaList).mkString
-      val parsed = parse(json)
+      if (followed.twitchUsers.isEmpty) stay using followed else {
 
-      val streaming = for {
-        JField("login", JString(login)) <- parsed
-        JField("up_time", JString(date)) <- parsed
-        JField("title", JString(title)) <- parsed
-      } yield (login.toString(), date.toString(), title.toString())
+        val commaList = followed.twitchUsers.foldLeft("") {
+          case (str, (key, value)) =>
+            str + key + ","
+        }.dropRight(1)
+        val url = "http://api.justin.tv/api/stream/list.json?channel="
+        println("url is" + url)
+        val json = Source.fromURL(url + commaList).mkString
+        val parsed = parse(json)
 
-      val updated = streaming.map {
-        case (login, date, title) =>
-          val twitchuser = followed.twitchUsers.get(login).get
-          if (!(twitchuser.lastAnnounced == date)) {
-            twitchuser.subscribers.foreach(a => a.actor ! Streaming(a.channel, login, title)) // Please don't kill me, Ornicar
-            twitchuser.name -> twitchuser.copy(lastAnnounced = date)
-          } else {
-            twitchuser.name -> twitchuser
-          }
-      } toMap
+        val streaming = for {
+          JField("login", JString(login)) <- parsed
+          JField("up_time", JString(date)) <- parsed
+          JField("title", JString(title)) <- parsed
+        } yield (login.toString(), date.toString(), title.toString())
 
-      stay using followed.copy(twitchUsers = updated) // placeholder
+        println("Streaming: " + streaming)
+
+        val updated = streaming.map {
+          case (login, date, title) =>
+            val twitchuser = followed.twitchUsers.get(login).get
+            if (!(twitchuser.lastAnnounced == date)) {
+              twitchuser.subscribers.foreach(a => a.actor ! Streaming(a.channel, login, title)) // Please don't kill me, Ornicar
+              twitchuser.name -> twitchuser.copy(lastAnnounced = date)
+            } else {
+              twitchuser.name -> twitchuser
+            }
+        } toMap
+
+        stay using followed.copy(twitchUsers = updated) // placeholder
+      }
 
     case Event(Subscribe(actor: ActorRef, channel: Channel, stream: String), followed: Followed) =>
       val subscriber = Subscriber(actor, channel, stream)
